pyproject.toml
[project]
name = "researchclaw"
version = "0.3.1"
description = "ResearchClaw — Autonomous Research Pipeline. Turn any research idea into a paper."
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "arxiv>=2.1",
    "numpy>=1.24",
]
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
anthropic = ["httpx>=0.24"]
web = ["scholarly>=1.7", "crawl4ai>=0.2", "tavily-python>=0.3"]
pdf = ["PyMuPDF>=1.23"]
all = [
    "httpx>=0.24",
    "scholarly>=1.7",
    "crawl4ai>=0.2",
    "tavily-python>=0.3",
    "PyMuPDF>=1.23",
    "huggingface-hub>=0.20",
    "matplotlib>=3.7",
    "scipy>=1.10",
]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "httpx>=0.24"]

[project.scripts]
researchclaw = "researchclaw.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["researchclaw", "sibyl", "arc"]

[tool.hatch.build.targets.wheel.force-include]
"researchclaw/templates/styles" = "researchclaw/templates/styles"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
